backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
pydantic==2.6.3
pydantic-settings==2.2.1
httpx==0.27.0
pytest==8.0.2
